refactor: implement cva on sphere groups and removed unecessary code
