--- src/stories/HeroSection/HeroSection.stories.ts
@@ -1,10 +1,0 @@
-import HeroSection from '@/components/HeroSection/HeroSection';
-
-export default {
-  title: 'HeroSection',
-  component: HeroSection,
-};
-
-export const Main = {
-  args: {},
-};
