+++ src/stories/ui/BallFrame.stories.ts
@@ -3,8 +3,16 @@ import BallFrame from '@/components/ui/BallFrame';
 export default {
   title: 'BallFrame',
   component: BallFrame,
+  args: {
+    hero: {
+      control: { type: 'select' },
+      options: ['true', 'false'],
+    },
+  },
 };
 
-export const Main = {
-  args: {},
+export const Frame = {
+  args: {
+    hero: 'true',
+  },
 };
